chore: update 'bundle-source' package to use AVA for testing
