+++ packages/bundle-source/test/test-circular.js
@@ -1,4 +1,5 @@
 /* global Compartment */
+
 import '@agoric/install-ses';
 import test from 'ava';
 import bundleSource from '..';
@@ -8,20 +9,20 @@ function evaluate(src, endowments) {
   return c.evaluate(src);
 }
 
-test(`external require('fs')`, async t => {
-  t.plan(1);
-  const { source: src1 } = await bundleSource(
-    `${__dirname}/../demo/external-fs.js`,
+test('circular export', async t => {
+  const { source: src1, sourceMap: map1 } = await bundleSource(
+    `${__dirname}/../demo/circular/a.js`,
     'nestedEvaluate',
   );
 
-  const myRequire = mod => t.is(mod, 'fs', 'required fs module');
-
   const nestedEvaluate = src => {
     // console.log('========== evaluating', src);
-    return evaluate(src, { nestedEvaluate, require: myRequire });
+    return evaluate(src, { nestedEvaluate });
   };
   // console.log(src1);
-  const srcMap1 = `(${src1})`;
-  nestedEvaluate(srcMap1)();
+  const srcMap1 = `(${src1})\n${map1}`;
+  const ex1 = nestedEvaluate(srcMap1)();
+
+  // console.log(err.stack);
+  t.is(ex1.default, 'Foo', `circular export is Foo`);
 });
