+++ packages/bundle-source/test/test-comment.js
@@ -1,6 +1,6 @@
 /* global Compartment */
 import '@agoric/install-ses';
-import { test } from 'tape-promise/tape';
+import test from 'ava';
 import bundleSource from '..';
 
 function evaluate(src, endowments) {
@@ -9,77 +9,55 @@ function evaluate(src, endowments) {
 }
 
 test('trailing comment', async t => {
-  try {
-    const { source: src1 } = await bundleSource(
-      `${__dirname}/../demo/comments/trailing-comment.js`,
-      'nestedEvaluate',
-    );
+  const { source: src1 } = await bundleSource(
+    `${__dirname}/../demo/comments/trailing-comment.js`,
+    'nestedEvaluate',
+  );
 
-    const nestedEvaluate = src => {
-      // console.log('========== evaluating', src);
-      return evaluate(src, { nestedEvaluate });
-    };
-    // console.log(src1);
-    const srcMap1 = `(${src1})`;
-    const ex1 = nestedEvaluate(srcMap1)();
+  const nestedEvaluate = src => {
+    // console.log('========== evaluating', src);
+    return evaluate(src, { nestedEvaluate });
+  };
+  // console.log(src1);
+  const srcMap1 = `(${src1})`;
+  const ex1 = nestedEvaluate(srcMap1)();
 
-    // console.log(err.stack);
-    t.equals(
-      typeof ex1.buildRootObject,
-      'function',
-      `buildRootObject is exported`,
-    );
-  } catch (e) {
-    t.isNot(e, e, 'unexpected exception');
-  } finally {
-    t.end();
-  }
+  // console.log(err.stack);
+  t.is(typeof ex1.buildRootObject, 'function', `buildRootObject is exported`);
 });
 
 test('comment block opener', async t => {
-  try {
-    t.plan(1);
-    const { source: src1 } = await bundleSource(
-      `${__dirname}/../demo/comments/block-opener.js`,
-      'nestedEvaluate',
-    );
+  t.plan(1);
+  const { source: src1 } = await bundleSource(
+    `${__dirname}/../demo/comments/block-opener.js`,
+    'nestedEvaluate',
+  );
 
-    const success = () => t.pass('body runs correctly');
+  const success = () => t.pass('body runs correctly');
 
-    const nestedEvaluate = src => {
-      // console.log('========== evaluating', src);
-      return evaluate(src, { nestedEvaluate, success });
-    };
-    // console.log(src1);
-    const srcMap1 = `(${src1})`;
-    nestedEvaluate(srcMap1)();
-  } catch (e) {
-    t.isNot(e, e, 'unexpected exception');
-  } finally {
-    t.end();
-  }
+  const nestedEvaluate = src => {
+    // console.log('========== evaluating', src);
+    return evaluate(src, { nestedEvaluate, success });
+  };
+  // console.log(src1);
+  const srcMap1 = `(${src1})`;
+  nestedEvaluate(srcMap1)();
 });
 
 test('comment block closer', async t => {
-  try {
-    t.plan(1);
-    const { source: src1 } = await bundleSource(
-      `${__dirname}/../demo/comments/block-closer.js`,
-      'nestedEvaluate',
-    );
+  t.plan(1);
+  const { source: src1 } = await bundleSource(
+    `${__dirname}/../demo/comments/block-closer.js`,
+    'nestedEvaluate',
+  );
 
-    const success = () => t.pass('body runs correctly');
+  const success = () => t.pass('body runs correctly');
 
-    const nestedEvaluate = src => {
-      // console.log('========== evaluating', src);
-      return evaluate(src, { nestedEvaluate, success });
-    };
-    // console.log(src1);
-    const srcMap1 = `(${src1})`;
-    nestedEvaluate(srcMap1)();
-  } catch (e) {
-    t.isNot(e, e, 'unexpected exception');
-  } finally {
-    t.end();
-  }
+  const nestedEvaluate = src => {
+    // console.log('========== evaluating', src);
+    return evaluate(src, { nestedEvaluate, success });
+  };
+  // console.log(src1);
+  const srcMap1 = `(${src1})`;
+  nestedEvaluate(srcMap1)();
 });
